move google books api call to new file
also add new Book class
